+++ support/models/src/local/index.ts
@@ -30,6 +30,7 @@ import "./GroupKeyManagementOverrides.js";
 import "./GroupsOverrides.js";
 import "./hwadr.js";
 import "./IlluminanceMeasurementOverrides.js";
+import "./JointFabricDatastoreOverrides.js";
 import "./ipadr.js";
 import "./ipv4adr.js";
 import "./ipv6adr.js";